TailTalk: Gate Pcap usage behind ethertalk flag
Refactors our code base so only if the "ethertalk" feature flag is set
do we actually pull in the pcap crate. This makes it so our packages
published on GitHub which has this feature disabled does not actually
require libpcap to be bundled (for macOS / Linux) or installed (for Windows).

Additionally bumps our deps to the latest versions and some minor changes
associated with the crate bumps.
