+++ examples/afp-server/Cargo.toml
@@ -10,7 +10,7 @@ path = "main.rs"
 
 [dependencies]
 clap = { workspace = true } #unified
-tailtalk = { workspace = true } #unified
+tailtalk = { workspace = true, features = ["ethertalk"] } #unified
 tailtalk-packets = { workspace = true } #unified
 tokio = { workspace = true, features = ["io-util"] } #unified
 tracing = { workspace = true } #unified
