+++ tailtalk-gui/Cargo.toml
@@ -11,14 +11,14 @@ path = "main.rs"
 [features]
 default = ["tashtalk"]
 tashtalk = ["dep:serialport"]
-ethertalk = ["dep:if-addrs"]
+ethertalk = ["dep:if-addrs", "tailtalk/ethertalk"]
 
 [dependencies]
 anyhow = { workspace = true } #unified
-dirs = "5"
+dirs = "6"
 hfs-reader = { workspace = true } #unified
-if-addrs = { version = "0.13", optional = true }
-rfd = "0.15"
+if-addrs = { version = "0.15", optional = true }
+rfd = "0.17"
 serde = { version = "1", features = ["derive"] }
 serialport = { version = "4", optional = true }
 slint = "1"
@@ -26,7 +26,7 @@ stuffit = { git = "https://github.com/Arctessa/stuffit-rs", branch = "huffman-me
 tailtalk = { workspace = true } #unified
 tailtalk-packets = { workspace = true } #unified
 tokio = { workspace = true, features = ["io-util"] } #unified
-toml = "0.8"
+toml = "^1.1"
 tracing = { workspace = true } #unified
 tracing-appender = "0.2"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
