+++ tailtalk/src/ddp.rs
@@ -1,4 +1,4 @@
-use rand::Rng;
+use rand::RngExt;
 use std::{
     collections::HashMap,
     io::{self, Error},
